Pull `applyMigration(id, migration)` method down to record based repo
